compute.oncoscore.timeseries

perform the OncoScore time series analysis for a list of genes and data times

compute.oncoscore.timeseries(
  data,
  filter.threshold = 0,
  analysis.mode = "Log2",
  cutoff.threshold = 21.09,
  file = NULL
)

Arguments

data

input data as result of the function perform.query.timeseries

filter.threshold

threshold to filter for a minimum number of citations for the genes

analysis.mode

logaritmic scores to be computed, i.e., log10, log2, natural log or log5

cutoff.threshold

threshold to be used to asses the oncogenes

file

should I save the results to text files?

Value

the performed OncoScores time series analysis

Examples

data(query.timepoints)
compute.oncoscore.timeseries(query.timepoints)
#> ### Computing oncoscore for timepoint 2012/03/01 
#> ### Processing data
#> ### Computing frequencies scores 
#> ### Estimating oncogenes
#> ### Results:
#> 	 ASXL1 -> 77.19348 
#> 	 IDH1 -> 74.24489 
#> 	 IDH2 -> 64.1649 
#> 	 SETBP1 -> 34.9485 
#> 	 TET2 -> 74.90108 
#> ### Computing oncoscore for timepoint 2013/03/01 
#> ### Processing data
#> ### Computing frequencies scores 
#> ### Estimating oncogenes
#> ### Results:
#> 	 ASXL1 -> 76.31902 
#> 	 IDH1 -> 78.71551 
#> 	 IDH2 -> 69.99559 
#> 	 SETBP1 -> 46.4559 
#> 	 TET2 -> 73.89695 
#> ### Computing oncoscore for timepoint 2014/03/01 
#> ### Processing data
#> ### Computing frequencies scores 
#> ### Estimating oncogenes
#> ### Results:
#> 	 ASXL1 -> 77.39202 
#> 	 IDH1 -> 81.07946 
#> 	 IDH2 -> 73.46995 
#> 	 SETBP1 -> 64.97398 
#> 	 TET2 -> 70.44331 
#> ### Computing oncoscore for timepoint 2015/03/01 
#> ### Processing data
#> ### Computing frequencies scores 
#> ### Estimating oncogenes
#> ### Results:
#> 	 ASXL1 -> 77.49295 
#> 	 IDH1 -> 82.55032 
#> 	 IDH2 -> 75.58179 
#> 	 SETBP1 -> 63.80208 
#> 	 TET2 -> 69.91466 
#> ### Computing oncoscore for timepoint 2016/03/01 
#> ### Processing data
#> ### Computing frequencies scores 
#> ### Estimating oncogenes
#> ### Results:
#> 	 ASXL1 -> 77.8392 
#> 	 IDH1 -> 83.11346 
#> 	 IDH2 -> 76.75356 
#> 	 SETBP1 -> 65.556 
#> 	 TET2 -> 69.81125 
#> $`2012/03/01`
#>        CitationsGene CitationsGeneInCancer  PercCit    alpha   1/alpha
#> ASXL1             91                    83 91.20879 6.507795 0.1536619
#> IDH1             488                   408 83.60656 8.930737 0.1119728
#> IDH2             234                   172 73.50427 7.870365 0.1270589
#> SETBP1            10                     5 50.00000 3.321928 0.3010300
#> TET2             196                   169 86.22449 7.614710 0.1313248
#>        PercCit*1/alpha OncoScore Clustering
#> ASXL1        14.015315  77.19348          1
#> IDH1          9.361663  74.24489          1
#> IDH2          9.339373  64.16490          1
#> SETBP1       15.051500  34.94850          1
#> TET2         11.323411  74.90108          1
#> 
#> $`2013/03/01`
#>        CitationsGene CitationsGeneInCancer  PercCit    alpha   1/alpha
#> ASXL1            149                   132 88.59060 7.219169 0.1385201
#> IDH1             753                   662 87.91501 9.556506 0.1046408
#> IDH2             336                   267 79.46429 8.392317 0.1191566
#> SETBP1            18                    11 61.11111 4.169925 0.2398125
#> TET2             302                   254 84.10596 8.238405 0.1213827
#>        PercCit*1/alpha OncoScore Clustering
#> ASXL1        12.271580  76.31902          1
#> IDH1          9.199493  78.71551          1
#> IDH2          9.468694  69.99559          1
#> SETBP1       14.655206  46.45590          1
#> TET2         10.209010  73.89695          1
#> 
#> $`2014/03/01`
#>        CitationsGene CitationsGeneInCancer  PercCit    alpha   1/alpha
#> ASXL1            208                   185 88.94231 7.700440 0.1298627
#> IDH1            1002                   903 90.11976 9.968667 0.1003143
#> IDH2             439                   364 82.91572 8.778077 0.1139202
#> SETBP1            36                    29 80.55556 5.169925 0.1934264
#> TET2             430                   342 79.53488 8.748193 0.1143093
#>        PercCit*1/alpha OncoScore Clustering
#> ASXL1        11.550289  77.39202          1
#> IDH1          9.040302  81.07946          1
#> IDH2          9.445772  73.46995          1
#> SETBP1       15.581571  64.97398          1
#> TET2          9.091579  70.44331          1
#> 
#> $`2015/03/01`
#>        CitationsGene CitationsGeneInCancer  PercCit     alpha    1/alpha
#> ASXL1            283                   250 88.33922  8.144658 0.12277986
#> IDH1            1300                  1188 91.38462 10.344296 0.09667164
#> IDH2             550                   467 84.90909  9.103288 0.10985042
#> SETBP1            64                    49 76.56250  6.000000 0.16666667
#> TET2             576                   452 78.47222  9.169925 0.10905215
#>        PercCit*1/alpha OncoScore Clustering
#> ASXL1        10.846277  77.49295          1
#> IDH1          8.834300  82.55032          1
#> IDH2          9.327299  75.58179          1
#> SETBP1       12.760417  63.80208          1
#> TET2          8.557564  69.91466          1
#> 
#> $`2016/03/01`
#>        CitationsGene CitationsGeneInCancer  PercCit     alpha    1/alpha
#> ASXL1            350                   309 88.28571  8.451211 0.11832624
#> IDH1            1576                  1446 91.75127 10.622052 0.09414377
#> IDH2             648                   557 85.95679  9.339850 0.10706810
#> SETBP1            89                    69 77.52809  6.475733 0.15442266
#> TET2             715                   558 78.04196  9.481799 0.10546521
#>        PercCit*1/alpha OncoScore Clustering
#> ASXL1        10.446516  77.83920          1
#> IDH1          8.637810  83.11346          1
#> IDH2          9.203230  76.75356          1
#> SETBP1       11.972094  65.55600          1
#> TET2          8.230712  69.81125          1
#>