perform.query.timeseries

perforn the query to PubMed for the time series analysis

perform.query.timeseries(
  list.of.genes,
  list.of.datatimes,
  gene.num.limit = 100,
  timepoints.limit = 10,
  custom.search = NA
)

Arguments

list.of.genes

The list of genes to be used in the queries to PubMed

list.of.datatimes

The list of time points to be used in the queries to PubMed

gene.num.limit

A limit to the genes to be considered in the analysis; this is done to limit the number of queries to PubMed

timepoints.limit

A limit to the time points to be considered in the analysis; this is done to limit the number of queries to PubMed

custom.search

A custom set of keyworkds to be used when quering PubMed

Value

The frequencies of the genes in the cancer related documents and in all the documents retireved on PubMed at the specified time points

Examples

data(genes)
data(timepoints)